+++ CHANGELOG.md
@@ -13,6 +13,7 @@ Note: In this file, do not use the hard wrap in the middle of a sentence for com
 
 - improve: picker now auto-removes bad symlinks
 - improve: random initialization for divider solver
+- improve: tagger tag will sort files according to the last modified time
 
 ## [0.1.0-alpha46] - 2024-10-13